feat(unauth_lu_page): button AddNewResource render LoginDialog when no auth
